Move Functional inline scripts to DAOS project (#153)
