+++ vars/runTestFunctional.groovy
@@ -7,7 +7,7 @@
  *
  */
 
-def call(Map config = [:]) {
+void call(Map config = [:]) {
   /**
    * runTestFunctional step method
    *
@@ -38,79 +38,25 @@ def call(Map config = [:]) {
    *                        Default env.STAGE_NAME.
    */
 
-    def test_rpms = 'false'
-    if (config['test_rpms'] == "true") {
-        test_rpms = 'true'
+    if (!fileExists('ci/functional/test_main.sh')) {
+        return runTestFunctionalV1(config)
     }
-    def functional_test_script = '''test_tag=$(git show -s --format=%%B | sed -ne "/^Test-tag%s:/s/^.*: *//p")
-                                    if [ -z "$test_tag" ]; then
-                                        test_tag=%s
-                                    fi
-                                    tnodes=$(echo $NODELIST | cut -d ',' -f 1-%d)
-                                    first_node=${NODELIST%%%%,*}
-                                    clush -B -S -o '-i ci_key' -l root -w "${first_node}" "set -ex
-                                      systemctl start nfs-server.service
-                                      mkdir -p /export/share
-                                      chown jenkins /export/share
-                                      echo \\"/export/share ${NODELIST//,/(rw,no_root_squash) }(rw,no_root_squash)\\" > /etc/exports
-                                      exportfs -ra"
-                                    clush -B -S -o '-i ci_key' -l root -w "${tnodes}" \
-                                      "set -ex
-                                       for i in 0 1; do
-                                           if [ -e /sys/class/net/ib\\\$i ]; then
-                                               if ! ifconfig ib\\\$i | grep \"inet \"; then
-                                                 {
-                                                   echo \"Found interface ib\\\$i down after reboot on \\\$HOSTNAME\"
-                                                   systemctl status
-                                                   systemctl --failed
-                                                   journalctl -n 500
-                                                   ifconfig ib\\\$i
-                                                   cat /sys/class/net/ib\\\$i/mode
-                                                   ifup ib\\\$i
-                                                 } | mail -s \"Interface found down after reboot\" $OPERATIONS_EMAIL
-                                               fi
-                                           fi
-                                       done
-                                       if ! grep /mnt/share /proc/mounts; then
-                                           mkdir -p /mnt/share
-                                           mount $first_node:/export/share /mnt/share
-                                       fi
-                                       if ''' + test_rpms + '''; then
-                                           # remove the install/ dir to be sure we're testing from RPMs
-                                           rm -rf install/
-                                       fi"
-                                    trap 'clush -B -S -o "-i ci_key" -l root -w "${tnodes}" \
-                                          "set -x; umount /mnt/share"' EXIT
-                                    # set DAOS_TARGET_OVERSUBSCRIBE env here
-                                    export DAOS_TARGET_OVERSUBSCRIBE=1
-                                    rm -rf install/lib/daos/TESTING/ftest/avocado ./*_results.xml
-                                    mkdir -p install/lib/daos/TESTING/ftest/avocado/job-results
-                                    ftest_arg="%s"
-                                    if ''' + test_rpms + '''; then
-                                        ssh -i ci_key -l jenkins "${first_node}" "set -ex
-                                          DAOS_TEST_SHARED_DIR=\\$(mktemp -d -p /mnt/share/)
-                                          trap \\"rm -rf \\$DAOS_TEST_SHARED_DIR\\" EXIT
-                                          export DAOS_TEST_SHARED_DIR
-                                          export TEST_RPMS=true
-                                          export REMOTE_ACCT=jenkins
-                                          /usr/lib/daos/TESTING/ftest/ftest.sh \\"$test_tag\\" \\"$tnodes\\" \\"$ftest_arg\\""
-                                        # now collect up the logs and store them like non-RPM test does
-                                        mkdir -p install/lib/daos/TESTING/
-                                        # scp doesn't copy symlinks, it resolves them
-                                        ssh -i ci_key -l jenkins "${first_node}" tar -C /var/tmp/ -czf - ftest | tar -C install/lib/daos/TESTING/ -xzf -
-                                    else
-                                        ./ftest.sh "$test_tag" "$tnodes" "$ftest_arg"
-                                    fi'''
 
-    config['script'] = String.format(functional_test_script,
-                                     config['pragma_suffix'],
-                                     config['test_tag'],
-                                     config['node_count'],
-                                     config['ftest_arg'])
+    Boolean test_rpms = false
+    if (config['test_rpms'] == "true") {
+        test_rpms = true
+    }
+    config['script'] = "TEST_TAG=" + config['test_tag'] + ' ' +
+                       "FTEST_ARG=" + config['ftest_arg'] + ' ' +
+                       "PRAGMA_SUFFIX=" + config['pragma_suffix'] + ' ' +
+                       "NODE_COUNT=" + config['node_count'] + ' ' +
+                       "OPERATIONS_EMAIL=" + env.OPERATIONS_EMAIL + ' ' +
+                       "ci/functional/test_main.sh"
+                                     
     config['junit_files'] = "install/lib/daos/TESTING/ftest/avocado/job-results/job-*/*.xml install/lib/daos/TESTING/ftest/*_results.xml"
     config['failure_artifacts'] = 'Functional'
 
-    if (test_rpms == 'true' && config['stashes']){
+    if (test_rpms && config['stashes']){
         // we don't need (and might not even have) stashes if testing
         // from RPMs
         config.remove('stashes')
